+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - The rack_test driver now fully supports the accessible name calculations
 - The rack_test driver now fully supports resolving accessible roles
+- Added an `accessible_description` `Node` method and filter
+- Deprecated the `described_by` filter in favour of `accessible_name`
+- The accessible description calculation now supports `aria-description` and has been handling of hidden text
 - Add `:test_id` selector
 - The `aria` filter will now filter for not having the attribute if the hash value is `nil`